♻️ [refactor] modify Camera Button #7
